"8:35am. Right now I am thinking about what my next move should be. Here are my thoughts.
* I cannot move to Maze until I stabilize the standard nets on one card poker.
* I still haven't even mastered the linear case.
* I think I should try Q learning with Zap again. I actually did not do this last time because embedding the actions was such a pain. But now I know better. I could probably do without embedding the actions.
* I do not think that using the scale to pare down the gradients coming from the head is wrong in deep nets.
* Hence, if only the head's weights should contain value estimates, then backprop is wrong. It is impossible to avoid propagating the global signal that contains the value aspect with just the symmetric weights.
* It is said that Q learning is unstable with non-linear approximation, but I am betting the same would apply with multilayer linear nets. I should ask about this.

Everything is going wrong for me. What did I even do in the last six months?

What is my job here? This is something I've been asking myself over and over during the night.

It is to correct the mistakes of optimization algorithms. That is the conclusion I've come to.

9am. My blood is boiling over this.

My near term goal will be to master the linear case. I will also forget about actor-critics for the time being. And for the deep case, I am giving up on backpropagation. I need to come up with something better. At very least, I will aim to avoid propagating the scale of the values in the gradient. That should stabilize things to a significant degree.

Right now in late 2018 there isn't an explanation for why Q learning is unstable with deep nets, but if one thinks about it, then it should be obvious - it is because of the way credit assignment is done in deep nets. The fact that the brain does not have symmetric weights is not an accident. It is to prevent runaway feedback that comes from updating value functions.

Forget model-based RL and metalearning. What I need to do first is take care of basic credit assignment issues in RL.

I am actually pretty sure about the scale aspect, but I am unsure how to propagate the gradients. Should I weight norm the weights? Or should I do something else?

9:35am. Well, for now, let me finish the talk by Joshua Bengio that I started yesterday and then I will go through the AGREL paper. Yesterday I read it and today I will seriously study it and the related references. They are going from the notion of biological plausibility, but I am going from the notion that backprop is simply wrong.

When it is time to restart, I will do the ML library from scratch again and start it off by making a Zap Q agent with optimistic init. I will make sure of that speedup. This time I will implement Nelsa or Polsa as well.

Then I will look into doing deep representations for RL properly."