I am just chilling right now.
Since there is not much I want to do, I guess I have no choice but to start programming. I'll start with the new matmult and make my way from there.

It won't be difficult to make the new actor critic. It will take me a day or two at most and after that I'll be done with it.

I'll finally be able to move from this whole scaling issue. It feels like I've been stuck on it forever.

Let me finish my reading and then I'll start.