"10:35am.
```
    inl fwd streams data weight s =
        inl out =
            inl data = basic data
            inl weight = basic weight
            mm data weight s

        wrap_split

    inl matmult_ac {data weight front streams} s =
        inl {out with policy value scale} = fwd streams data weight s // trace is part of policy
        wait_on streams s
        {
        out
        bck=inl _ ->
            bck streams weight data policy s
            bck streams weight {rescaled={data scale}} value s
            bck streams weight {blocked=data} scale s
            update_covariance front data s
            wait_on streams s
        }
```

There are so many complications here. In addition to making the view the default, I really need primal and adjoint to be a part of the tensor. If I have to write Struct.map' one more time...

10:40am. No, I can't do it.

Not like now and like this.

Now while the failure of my last few experiments looms fresh in my mind. What I need to do is do this inefficiently and to the point.

Yeah, I am starting to be aware that the design of the current library is broken...again, but embarking on a week long redesign is something I am not up for right now.

I am going to do this as straightforwardly as possible within the confines of the current design even if it will be inefficient."