"12:25pm. Done with breakfast. Let me go help my brother out.
4:55pm. I'll admit I did not expect this helping out would involve 4.5 hours of backbreaking labor. Ugh, I am not sure I'll be able to continue right now. I need a break at this point.

https://arxiv.org/abs/1711.02257
GradNorm: Gradient Normalization for Adaptive Loss Balancing in Deep Multitask Networks

I feel like reading a bit. While I was working, my mind was wondering to what I will do after this experiment. What I should be doing is maximizing both noise and stability in the network. KFAC won't cut it. It needs to be more. Multilayer nets need so much more to be stable. They need to be stable regardless of initialization."