Yeah, not going to be doing anymore programming.
I am really engrossed with gradient normalization right now. Making it input dependent is definitely the way to go. There is just so much that is possible with good optimizers. I'd like to know how to do input dependent decorrelation of gradients.

I think in a RL setting, I should even consider splitting the backward passes into two - one for the value and one for the policy part. The reason to do that would be to apply KFAC using two separate decorrelation matrices on the back side.

I think that could be pretty powerful for multitask learning which RL is.

Assuming I get good results with this current experiment, I should definitely consider digging deeper in this direction.

...Actually...on reflection that should not be necessary. If I do a proper job in the head then thing will be equivalent in the bottom layers. No need for two passes. This is because of the linearity of the backward pass. I'd just be doubling the work for no reason.

It would only make sense if I am not doing input dependent gradient normalization in the head.

I'll cut this rant short and call it a day. I want to have some fun.