I am finally up. I spent way too long in bed.
I think I will finally do some programming. I had some time to think about how I want to do this yesterday. No doubt when all this is over I am going to do a redesign of the ML library assuming I can get this to work as it should.

I am trying to reason out whether input dependent normalization would be a good idea in the hidden layers and am not making any progress. I think it should stabilize gradients, but I cannot piece together even whether it should work at all.

One thing all of this is demonstrating to me is that if I want to get full efficiency, I am going to seriously have to upgrade the ML library.

I think I could go a lot further with the tree view tensors. They are the key to doing everything properly.

But forget them for now.

I need to get deep RL to work. To do that, I need to make reward rescaling automatic. That will give the absolute biggest benefit to start with.

If this fails, I might seriously have to consider giving up on deep RL. The experiment is that important.

In order to draw out the true power of deep nets, the nets need to be stabilized to the limit and then I have to noise them to the limit. That will purge the inefficiencies from them.

But if the normalization idea fails, then I will be in a trully difficult situation.

It is the decisive moment. If I can push through here, then I can claim everything. If not...

Let me go once more from the top, I need to see where I was yesterday...ah, yeah, I decided I am not going to track the variance, but rather the absolute magnitudes of the error.

Let me deal with that first.