Let me finally resume this. I had some work on the side that distracted me, but thankfully that should be done.
Let me refocus. What is my goal?

First of all, The RL agent should be invariant to reward scaling assuming the scale's bias is proportional to the reward's scale.

That is something that I did not get while I was trying out the Gaussian KL divergence. If I can get that then I will have met the first prerequisite.

In fact, I am saying this merely because if this does not happen, there is some bug in my code.

Let me do this and then I will bring in the RNN.

No doubt all the inefficiencies that are now present in the AC will slow it down to a crawl, but I'll worry about that after I get the experiment to succeed. It is time to go.