"2:35pm.
```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 242 and 758 out of 1000.
The time was 00:00:03.8009620 for: iteration 0
Starting timing for: iteration 1
Winrate is 337 and 663 out of 1000.
The time was 00:00:04.7343228 for: iteration 1
Starting timing for: iteration 2
Winrate is 414 and 586 out of 1000.
The time was 00:00:05.5053677 for: iteration 2
Starting timing for: iteration 3
Winrate is 483 and 517 out of 1000.
The time was 00:00:05.7014451 for: iteration 3
Starting timing for: iteration 4
Winrate is 539 and 461 out of 1000.
The time was 00:00:05.9702669 for: iteration 4
Starting timing for: iteration 5
Winrate is 560 and 440 out of 1000.
The time was 00:00:06.1836439 for: iteration 5
Starting timing for: iteration 6
Winrate is 591 and 409 out of 1000.
The time was 00:00:05.9966316 for: iteration 6
Starting timing for: iteration 7
Winrate is 600 and 400 out of 1000.
The time was 00:00:05.7427348 for: iteration 7
Starting timing for: iteration 8
Winrate is 637 and 363 out of 1000.
The time was 00:00:05.7306378 for: iteration 8
Starting timing for: iteration 9
Winrate is 631 and 369 out of 1000.
The time was 00:00:05.9456384 for: iteration 9
Starting timing for: iteration 10
Winrate is 661 and 339 out of 1000.
The time was 00:00:05.6813703 for: iteration 10
Starting timing for: iteration 11
Winrate is 682 and 318 out of 1000.
The time was 00:00:05.5308817 for: iteration 11
Starting timing for: iteration 12
Winrate is 650 and 350 out of 1000.
The time was 00:00:05.7260524 for: iteration 12
Starting timing for: iteration 13
Winrate is 661 and 339 out of 1000.
The time was 00:00:05.6235724 for: iteration 13
Starting timing for: iteration 14
Winrate is 670 and 330 out of 1000.
The time was 00:00:05.6036077 for: iteration 14
```

I had to divide the reward by 10 to get it to work with this LR, but this is what I get currently. Let me bring back learned scaling.

2:35pm.

```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 311 and 689 out of 1000.
The time was 00:00:02.5945731 for: iteration 0
Starting timing for: iteration 1
Winrate is 301 and 699 out of 1000.
The time was 00:00:02.9040031 for: iteration 1
Starting timing for: iteration 2
Winrate is 361 and 639 out of 1000.
The time was 00:00:03.0761655 for: iteration 2
Starting timing for: iteration 3
Winrate is 377 and 623 out of 1000.
The time was 00:00:03.1701412 for: iteration 3
Starting timing for: iteration 4
Winrate is 351 and 649 out of 1000.
The time was 00:00:03.2768093 for: iteration 4
Starting timing for: iteration 5
Winrate is 407 and 593 out of 1000.
The time was 00:00:03.2992288 for: iteration 5
Starting timing for: iteration 6
Winrate is 389 and 611 out of 1000.
The time was 00:00:03.2100255 for: iteration 6
Starting timing for: iteration 7
Winrate is 402 and 598 out of 1000.
The time was 00:00:03.3390394 for: iteration 7
Starting timing for: iteration 8
Winrate is 421 and 579 out of 1000.
The time was 00:00:03.4805438 for: iteration 8
Starting timing for: iteration 9
Winrate is 477 and 523 out of 1000.
The time was 00:00:03.9890622 for: iteration 9
Starting timing for: iteration 10
Winrate is 446 and 554 out of 1000.
The time was 00:00:04.4786990 for: iteration 10
Starting timing for: iteration 11
Winrate is 512 and 488 out of 1000.
The time was 00:00:04.5120347 for: iteration 11
Starting timing for: iteration 12
Winrate is 483 and 517 out of 1000.
The time was 00:00:04.2414785 for: iteration 12
Starting timing for: iteration 13
Winrate is 516 and 484 out of 1000.
The time was 00:00:04.2020331 for: iteration 13
Starting timing for: iteration 14
Winrate is 507 and 493 out of 1000.
The time was 00:00:04.3802763 for: iteration 14
```

Ok, this is something. Let me remove the division of reward by 10 and increase the scale's bias by 10."