"2:40pm.
```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 313 and 687 out of 1000.
The time was 00:00:05.1688218 for: iteration 0
Starting timing for: iteration 1
Winrate is 475 and 525 out of 1000.
The time was 00:00:06.5683096 for: iteration 1
Starting timing for: iteration 2
Winrate is 575 and 425 out of 1000.
The time was 00:00:06.2588558 for: iteration 2
Starting timing for: iteration 3
Winrate is 599 and 401 out of 1000.
The time was 00:00:05.8412235 for: iteration 3
Starting timing for: iteration 4
Winrate is 556 and 444 out of 1000.
The time was 00:00:05.0471389 for: iteration 4
Starting timing for: iteration 5
Winrate is 526 and 474 out of 1000.
The time was 00:00:04.7809030 for: iteration 5
Starting timing for: iteration 6
Winrate is 552 and 448 out of 1000.
The time was 00:00:05.2805901 for: iteration 6
Starting timing for: iteration 7
Winrate is 606 and 394 out of 1000.
The time was 00:00:05.3529169 for: iteration 7
Starting timing for: iteration 8
Winrate is 593 and 407 out of 1000.
The time was 00:00:04.6571832 for: iteration 8
Starting timing for: iteration 9
Winrate is 593 and 407 out of 1000.
The time was 00:00:05.3068768 for: iteration 9
Starting timing for: iteration 10
Winrate is 613 and 387 out of 1000.
The time was 00:00:05.2019797 for: iteration 10
Starting timing for: iteration 11
Winrate is 622 and 378 out of 1000.
The time was 00:00:05.0641976 for: iteration 11
Starting timing for: iteration 12
Winrate is 589 and 411 out of 1000.
The time was 00:00:04.6465400 for: iteration 12
Starting timing for: iteration 13
Winrate is 576 and 424 out of 1000.
The time was 00:00:04.7832999 for: iteration 13
Starting timing for: iteration 14
Winrate is 564 and 436 out of 1000.
The time was 00:00:04.9694188 for: iteration 14
```

I keep hoping for spectacular performance, but I am not getting it. Let me divide the scale and reward by both 10.

2:45pm.

```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 312 and 688 out of 1000.
The time was 00:00:05.3205599 for: iteration 0
Starting timing for: iteration 1
Winrate is 358 and 642 out of 1000.
The time was 00:00:08.5361657 for: iteration 1
Starting timing for: iteration 2
Winrate is 516 and 484 out of 1000.
The time was 00:00:07.1737115 for: iteration 2
Starting timing for: iteration 3
Winrate is 533 and 467 out of 1000.
The time was 00:00:06.8958136 for: iteration 3
Starting timing for: iteration 4
Winrate is 424 and 576 out of 1000.
The time was 00:00:05.0207250 for: iteration 4
Starting timing for: iteration 5
Winrate is 345 and 655 out of 1000.
The time was 00:00:04.0262553 for: iteration 5
Starting timing for: iteration 6
Winrate is 414 and 586 out of 1000.
The time was 00:00:04.4824053 for: iteration 6
Starting timing for: iteration 7
Winrate is 401 and 599 out of 1000.
The time was 00:00:04.6144715 for: iteration 7
Starting timing for: iteration 8
Winrate is 390 and 610 out of 1000.
The time was 00:00:04.4572736 for: iteration 8
Starting timing for: iteration 9
Winrate is 393 and 607 out of 1000.
The time was 00:00:04.6247615 for: iteration 9
Starting timing for: iteration 10
Winrate is 449 and 551 out of 1000.
The time was 00:00:04.6595018 for: iteration 10
Starting timing for: iteration 11
Winrate is 435 and 565 out of 1000.
The time was 00:00:04.7613548 for: iteration 11
Starting timing for: iteration 12
Winrate is 448 and 552 out of 1000.
The time was 00:00:04.7118797 for: iteration 12
Starting timing for: iteration 13
Winrate is 450 and 550 out of 1000.
The time was 00:00:04.5768740 for: iteration 13
Starting timing for: iteration 14
Winrate is 415 and 585 out of 1000.
The time was 00:00:04.8469624 for: iteration 14
```

Aghhh...

Let me remove that `+ epsilon -10`. The two should be perfectly equal when rewards are divided."