``` The learning_rate is 2 ** -8 ------ The CudaRandom pseudorandom seed is 0 Starting timing for: iteration 0 Winrate is 309 and 691 out of 1000. The time was 00:00:04.8054639 for: iteration 0 Starting timing for: iteration 1 Winrate is 419 and 581 out of 1000. The time was 00:00:08.8283825 for: iteration 1 Starting timing for: iteration 2 Winrate is 539 and 461 out of 1000. The time was 00:00:09.3546909 for: iteration 2 Starting timing for: iteration 3 Winrate is 584 and 416 out of 1000. The time was 00:00:08.5319208 for: iteration 3 Starting timing for: iteration 4 Winrate is 623 and 377 out of 1000. The time was 00:00:09.1343199 for: iteration 4 ```
This is with a reward scale of 10. Let me try 100 and 1.

```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 325 and 675 out of 1000.
The time was 00:00:05.2906932 for: iteration 0
Starting timing for: iteration 1
Winrate is 505 and 495 out of 1000.
The time was 00:00:07.4908606 for: iteration 1
Starting timing for: iteration 2
Winrate is 461 and 539 out of 1000.
The time was 00:00:07.3381072 for: iteration 2
Starting timing for: iteration 3
Winrate is 510 and 490 out of 1000.
The time was 00:00:08.1880738 for: iteration 3
Starting timing for: iteration 4
Winrate is 519 and 481 out of 1000.
The time was 00:00:09.0005853 for: iteration 4
```

This is 100.

Let me try 1.

```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 330 and 670 out of 1000.
The time was 00:00:05.6383388 for: iteration 0
Starting timing for: iteration 1
Winrate is 441 and 559 out of 1000.
The time was 00:00:07.6285423 for: iteration 1
Starting timing for: iteration 2
Winrate is 548 and 452 out of 1000.
The time was 00:00:08.9215921 for: iteration 2
Starting timing for: iteration 3
Winrate is 628 and 372 out of 1000.
The time was 00:00:08.8819501 for: iteration 3
Starting timing for: iteration 4
Winrate is 629 and 371 out of 1000.
The time was 00:00:08.5500731 for: iteration 4
```

This the power of noise huh? I hate how different the runs are.

Let me increase just the scale by 10x.