"3:20pm.
```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 294 and 706 out of 1000.
The time was 00:00:05.0818511 for: iteration 0
Starting timing for: iteration 1
Winrate is 209 and 791 out of 1000.
The time was 00:00:08.5371084 for: iteration 1
Starting timing for: iteration 2
Winrate is 245 and 755 out of 1000.
The time was 00:00:09.1660426 for: iteration 2
Starting timing for: iteration 3
Winrate is 346 and 654 out of 1000.
The time was 00:00:08.7154345 for: iteration 3
Starting timing for: iteration 4
Winrate is 449 and 551 out of 1000.
The time was 00:00:09.1513221 for: iteration 4
```

It takes it a surprisingly long time to correct the bias.

3:25pm.

```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 344 and 656 out of 1000.
The time was 00:00:04.8675672 for: iteration 0
Starting timing for: iteration 1
Winrate is 468 and 532 out of 1000.
The time was 00:00:07.4398556 for: iteration 1
Starting timing for: iteration 2
Winrate is 544 and 456 out of 1000.
The time was 00:00:08.3059595 for: iteration 2
Starting timing for: iteration 3
Winrate is 666 and 334 out of 1000.
The time was 00:00:08.0002349 for: iteration 3
Starting timing for: iteration 4
Winrate is 691 and 309 out of 1000.
The time was 00:00:07.8867973 for: iteration 4
```

This is with reward scale = 1000. What the hell? So it is not taking a long time to correct the bias?

```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 315 and 685 out of 1000.
The time was 00:00:04.8875918 for: iteration 0
Starting timing for: iteration 1
Winrate is 418 and 582 out of 1000.
The time was 00:00:05.4986115 for: iteration 1
Starting timing for: iteration 2
Winrate is 500 and 500 out of 1000.
The time was 00:00:05.8513570 for: iteration 2
Starting timing for: iteration 3
Winrate is 523 and 477 out of 1000.
The time was 00:00:06.1567501 for: iteration 3
Starting timing for: iteration 4
Winrate is 547 and 453 out of 1000.
The time was 00:00:05.9723237 for: iteration 4
```

This is with 1 / 1000. So it is not taking a long time to correct bias. Maybe it just took a bad beat that time?

3:30pm. You know, since this is the ideal chance to do it, let me put KFAC just into the policy and the trace part.

...No wait, the trace. Yeah, what do I about the trace?

For now, how about I disable it?

By all means, the gradients for it should be rescaled by reward as well, but I can't afford that at the moment.

Let me just plug in KFAC for now.

3:35pm.

```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 377 and 623 out of 1000.
The time was 00:00:05.1143949 for: iteration 0
Starting timing for: iteration 1
Winrate is 523 and 477 out of 1000.
The time was 00:00:04.8630263 for: iteration 1
Starting timing for: iteration 2
Winrate is 538 and 462 out of 1000.
The time was 00:00:04.6265791 for: iteration 2
Starting timing for: iteration 3
Winrate is 570 and 430 out of 1000.
The time was 00:00:04.6593159 for: iteration 3
Starting timing for: iteration 4
Winrate is 569 and 431 out of 1000.
The time was 00:00:04.9781217 for: iteration 4
```

KFAC's back pass stabilizes things to a significant degree for some reason. With it, every run is like this now.

Let me remove it to make sure that I am not imagining things. This might be due to the trace.

3:40pm.

```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 319 and 681 out of 1000.
The time was 00:00:04.4131350 for: iteration 0
Starting timing for: iteration 1
Winrate is 473 and 527 out of 1000.
The time was 00:00:05.4986221 for: iteration 1
Starting timing for: iteration 2
Winrate is 549 and 451 out of 1000.
The time was 00:00:05.6285137 for: iteration 2
Starting timing for: iteration 3
Winrate is 428 and 572 out of 1000.
The time was 00:00:06.3937073 for: iteration 3
Starting timing for: iteration 4
Winrate is 437 and 563 out of 1000.
The time was 00:00:07.7907592 for: iteration 4
```

This is how it does without KFAC's back pass. Let me fiddle with reward scale so I can see if there are any differences.

```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 319 and 681 out of 1000.
The time was 00:00:04.2977557 for: iteration 0
Starting timing for: iteration 1
Winrate is 473 and 527 out of 1000.
The time was 00:00:05.6807410 for: iteration 1
Starting timing for: iteration 2
Winrate is 549 and 451 out of 1000.
The time was 00:00:05.6236217 for: iteration 2
Starting timing for: iteration 3
Winrate is 428 and 572 out of 1000.
The time was 00:00:06.4114018 for: iteration 3
Starting timing for: iteration 4
Winrate is 437 and 563 out of 1000.
The time was 00:00:07.5854268 for: iteration 4
```

Yes, the variability in results was due to the trace. Damn that thing.

3:40pm. But essentially, the experiment is a success.

I finally have the reward invariance that I've been seeking. I need to play around more with learning rates. Let me do so.

```
The learning_rate is 2 ** -7
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 190 and 810 out of 1000.
The time was 00:00:08.9043610 for: iteration 0
Starting timing for: iteration 1
Winrate is 265 and 735 out of 1000.
The time was 00:00:11.9800060 for: iteration 1
Starting timing for: iteration 2
Winrate is 356 and 644 out of 1000.
The time was 00:00:12.5577852 for: iteration 2
Starting timing for: iteration 3
Winrate is 456 and 544 out of 1000.
The time was 00:00:11.3169500 for: iteration 3
Starting timing for: iteration 4
Winrate is 504 and 496 out of 1000.
The time was 00:00:10.0524410 for: iteration 4
```

Damn, the thing sure is slow. Let me lower the lr by 2.

```
The learning_rate is 2 ** -9
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 278 and 722 out of 1000.
The time was 00:00:03.9162733 for: iteration 0
Starting timing for: iteration 1
Winrate is 369 and 631 out of 1000.
The time was 00:00:05.7698644 for: iteration 1
Starting timing for: iteration 2
Winrate is 482 and 518 out of 1000.
The time was 00:00:06.3419334 for: iteration 2
Starting timing for: iteration 3
Winrate is 546 and 454 out of 1000.
The time was 00:00:07.1210694 for: iteration 3
Starting timing for: iteration 4
Winrate is 432 and 568 out of 1000.
The time was 00:00:04.5630618 for: iteration 4
```

Let me go for 15. I want to see how far it can go normally. Then I will put KFAC back in."