"4:30pm. Let me try once again, this time without TD. I'll set the trace to 1.
```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 428 and 572 out of 1000.
The time was 00:00:05.6894482 for: iteration 0
Starting timing for: iteration 1
Winrate is 655 and 345 out of 1000.
The time was 00:00:06.0850823 for: iteration 1
Starting timing for: iteration 2
Winrate is 694 and 306 out of 1000.
The time was 00:00:06.0093424 for: iteration 2
Starting timing for: iteration 3
Winrate is 709 and 291 out of 1000.
The time was 00:00:06.0501187 for: iteration 3
Starting timing for: iteration 4
Winrate is 680 and 320 out of 1000.
The time was 00:00:05.7786707 for: iteration 4
Starting timing for: iteration 5
Winrate is 717 and 283 out of 1000.
The time was 00:00:05.9675778 for: iteration 5
Starting timing for: iteration 6
Winrate is 725 and 275 out of 1000.
The time was 00:00:05.7427718 for: iteration 6
Starting timing for: iteration 7
Winrate is 697 and 303 out of 1000.
The time was 00:00:05.8675692 for: iteration 7
```

Ok, so it not completely hopeless. So it seems TD 0.5 leads to quite a lot of bias when combined with scaling.

4:35pm.

```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 476 and 524 out of 1000.
The time was 00:00:05.7479973 for: iteration 0
Starting timing for: iteration 1
Winrate is 620 and 380 out of 1000.
The time was 00:00:06.0342971 for: iteration 1
Starting timing for: iteration 2
Winrate is 694 and 306 out of 1000.
The time was 00:00:06.6983102 for: iteration 2
Starting timing for: iteration 3
Winrate is 688 and 312 out of 1000.
The time was 00:00:07.0207198 for: iteration 3
Starting timing for: iteration 4
Winrate is 701 and 299 out of 1000.
The time was 00:00:06.4966565 for: iteration 4
Starting timing for: iteration 5
Winrate is 713 and 287 out of 1000.
The time was 00:00:06.9100308 for: iteration 5
Starting timing for: iteration 6
Winrate is 740 and 260 out of 1000.
The time was 00:00:05.8679094 for: iteration 6
Starting timing for: iteration 7
Winrate is 707 and 293 out of 1000.
The time was 00:00:06.0564467 for: iteration 7
Starting timing for: iteration 8
Winrate is 696 and 304 out of 1000.
The time was 00:00:06.1198802 for: iteration 8
Starting timing for: iteration 9
Winrate is 689 and 311 out of 1000.
The time was 00:00:06.2978315 for: iteration 9
Starting timing for: iteration 10
Winrate is 694 and 306 out of 1000.
The time was 00:00:06.0104598 for: iteration 10
Starting timing for: iteration 11
Winrate is 670 and 330 out of 1000.
The time was 00:00:06.0207554 for: iteration 11
Starting timing for: iteration 12
Winrate is 693 and 307 out of 1000.
The time was 00:00:06.3430244 for: iteration 12
Starting timing for: iteration 13
Winrate is 692 and 308 out of 1000.
The time was 00:00:06.4227459 for: iteration 13
Starting timing for: iteration 14
Winrate is 713 and 287 out of 1000.
The time was 00:00:06.7596870 for: iteration 14
```

```
        inm trace = half

        inm R = r + discount * (trace * R' + (one - trace) * primal value')
        inm error = R - value

        inm _ = sqr (abs (primal error) - scale) / num 4 |> as_cost
        inm _ = sqr error / num 4 |> as_cost
```

I am messing with the learning speed of the scale and value nets. Let me set the trace to zero.

4:40pm. It does very poorly with td 0 and reward scaling. But TD 0 works excellently on its own.

4:50pm.

```
The learning_rate is 2 ** -9
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 330 and 670 out of 1000.
The time was 00:00:05.2634994 for: iteration 0
Starting timing for: iteration 1
Winrate is 493 and 507 out of 1000.
The time was 00:00:07.6624055 for: iteration 1
Starting timing for: iteration 2
Winrate is 615 and 385 out of 1000.
The time was 00:00:08.6496878 for: iteration 2
Starting timing for: iteration 3
Winrate is 692 and 308 out of 1000.
The time was 00:00:07.9477706 for: iteration 3
Starting timing for: iteration 4
Winrate is 709 and 291 out of 1000.
The time was 00:00:08.3893457 for: iteration 4
Starting timing for: iteration 5
Winrate is 731 and 269 out of 1000.
The time was 00:00:08.1763502 for: iteration 5
Starting timing for: iteration 6
Winrate is 716 and 284 out of 1000.
The time was 00:00:07.7659688 for: iteration 6
Starting timing for: iteration 7
Winrate is 742 and 258 out of 1000.
The time was 00:00:07.1013889 for: iteration 7
Starting timing for: iteration 8
Winrate is 752 and 248 out of 1000.
The time was 00:00:06.2998661 for: iteration 8
Starting timing for: iteration 9
Winrate is 709 and 291 out of 1000.
The time was 00:00:06.4969849 for: iteration 9
Starting timing for: iteration 10
Winrate is 671 and 329 out of 1000.
The time was 00:00:05.7438780 for: iteration 10
Starting timing for: iteration 11
Winrate is 690 and 310 out of 1000.
The time was 00:00:06.0340920 for: iteration 11
Starting timing for: iteration 12
Winrate is 669 and 331 out of 1000.
The time was 00:00:05.1685594 for: iteration 12
Starting timing for: iteration 13
Winrate is 651 and 349 out of 1000.
The time was 00:00:05.1165526 for: iteration 13
Starting timing for: iteration 14
Winrate is 649 and 351 out of 1000.
The time was 00:00:05.2100255 for: iteration 14
```

This is with TD 0.

It did really well for a bit, but then it got knocked off course. I just needed to lower the LR.

Let me add an epsilon to reward scaling.

```
inm scaled_error = error / (min scale (epsilon -3)) // Is used as reward for the actor.
```

Let me do it like this actually."