"4:55pm. It works horribly.
5pm. Hmmmmmm...

Basically I cannot estimate the true worth of this idea. I've been fiddling with it for how many hours now?

Here is what I am going to do...

5:20pm. What I am going to do is forget about RL for a while and put the idea into the RNN. I'll normalize the gradients on the char RNN task and see where that gets me. If that makes the performance better over the baseline, then I know I have something. If it does not, then I am just being deluded.

If that turns out to be the case, I am going to throw in the towel on deep RL. Simply put, it works well, but only when I set the hyperparameters pretty much exactly. Such a brittle algorithm cannot possibly be my workhorse.

What I would need to do in this case if I intend to use it despite that would be to automate all the hyperparameter tunning somehow.

5:30pm. I am not sure I feel like doing this right now.

What I have to do is easy enough, but it would take some time to implement and test.

5:35pm. There is no making up for lost time. What I so desperately want is not coming into my hands, but at least unlike last time I have Spiral. No matter how things end up, I will not turn my back on programming again. Neither will I turn my back on my talent.

Though my ideal is laziness, I will work hard with the dream of automating everything at some point.

If this is not enough, then I simply need more knowledge and more training.

I am going to stop for the day here. I did well despite the results not going my way and the distractions.

Tomorrow I will give it my all to make the gradient normalization idea work on RNNs."