It is time to start the working day.
I am going to make a dirty impelmentation of input dependent gradient normalization in the vanilla RNN.

Since the library needs a redesign anyway, I am not going to worry about elegance. I need to remove a lot of stuff and redo a lot of stuff anyway.

Let me just do this so I can put it out of my mind and go forward with a clear head.