"10:05am. The idea is a failure.
```
scale = weightf {init dim=outer,1}
```

Let me try it like this.

10:15am.

```
[|
    [|NaN|]
    [|NaN|]
    [|NaN|]
```

I am printing out the scale and noticing it gets nan. What the hell?

10:15am.

```
[|
    [|-233.2917|]
    [|-233.2684|]
    [|-233.7439|]
    [|-232.4382|]
```

In the first place why is it going in the negative direction? I added that abs as a safeguard if it ever went to very low values. There is something seriously flawed in my implementation of it."