"10:35am. Nope.
Let me once again put the full dimension.

```
    [|-7.070951; -2.903048; -12.30885; -23.84941; -28.35566; 30.51403; 7.636234; 8.338362; -10.81547; -15.5
```

Yeah, now it is more stable, but...all of these should be zero.

```
    inl grad_rescale {scale input} =
        inm error = (primal input) - scale
        inm cost = abs error / num 2 |> as_cost
        //inm _ = grad_rescale {scale input}
        succ error
```

Let me try the `abs` error. I am starting to see a pattern here.

I've been prepared for the method not to work, but the way it is not working here is going way beyond my expectations. This is beyond crazy. So this is gradient descent at work at learning the average. I do not think there is a error in the code. Gradient descent really is blowing up like a donkey here."