"11:30am. No I am not. Let me put full KFAC on scale. Right now it just blows up like a monkey.
11:35am.

```
[|0.02184281; 0.2302872; 0.01069842; -0.01563747; 1.061488; 0.006848854; 0.05572461; 0.4098176; 0.8330441; 0.3422448; -0.006674376; 0.4086649; 0.0381816; 0.4051294; 0.02239412; 0.6400268; 0.5634511; 0.5512532; 0.0137551; 0.7754208; 0.6875938; 0.123381; 2.702972; 1.22917; 0.06975741; 1.097999; 0.1465657; 0.06274902; 0.01514213; 0.01353322; -0.04716377; -0.05186953; 0.6522677; 0.008116579; 1.565968; 2.015915; -0.07137185; 0.7810962; 0.752734; 0.8173114; 0.05934986; 0.1173418; -0.1089733; 0.4892181; 0.025905; 1.034765; 0.04670846; 1.941355; 0.1585053; 0.03289099; 0.04194008; 0.7968829; 0.3699775; 0.1906622; -0.05228284; 0.767042; 0.2237523; 0.009242455; 2.173566; 2.138093; -0.04193718; 0.01063335; 0.04835882; 0.006957147; 0.02034717; 0.3798118; 0.001252593; 0.03307566; 0.5120092; 0.3734631; 1.902908; 0.7422214; -0.1317578; 0.7668421; -0.0008208924; 0.02798027; 0.2310873; 0.6418211; 1.338976; 0.4793471; 0.01233385; 0.8177496; 0.2806926; 0.6492403; 1.002108; 0.08280841; 0.04496405; 0.08170664; 0.08038049; 0.03107221; 0.03050445; 0.6575188; 1.063542; 0.3574717; 0.1658798; 0.08805285; 0.1163902; 0.02096236; 0.001253911; -0.113983; 0.03585254; 0.02264342; 0.9510776; 0.5036459...|]
```

Normalizing the gradients for scale make it stop blowing up.

All in all, this is quite a powerful lesson for me to swallow. I am in awe. Despite it not blowing up, it is still quite wrong in its results.

I got sidetracked. Let me finally test what I wanted, which is normalize the gradients and then I will stop for a bit.

11:45am.

```
The learning rate is 2 ** -13
Starting timing for: iteration 0
The time was 00:00:07.2860088 for: iteration 0
Training: 3.15674458779669
The learning rate is 2 ** -13
Starting timing for: iteration 1
The time was 00:00:07.2184873 for: iteration 1
Training: 2.57054204174646
The learning rate is 2 ** -13
Starting timing for: iteration 2
The time was 00:00:07.5795513 for: iteration 2
Training: 2.2733690627878
The learning rate is 2 ** -13
Starting timing for: iteration 3
The time was 00:00:07.1043044 for: iteration 3
Training: 2.15302204636528
The learning rate is 2 ** -13
Starting timing for: iteration 4
The time was 00:00:07.0105341 for: iteration 4
Training: 2.04492749152568
```

Hmmm, is this better than usual?

I am not using KFAC for the main weights.

...No difference at all. In fact it is suspicious that there is no difference at all.

```
The learning rate is 2 ** -13
Starting timing for: iteration 0
The time was 00:00:07.4386708 for: iteration 0
Training: 3.9739956214257
The learning rate is 2 ** -13
Starting timing for: iteration 1
The time was 00:00:07.3476784 for: iteration 1
Training: 3.51140631627127
The learning rate is 2 ** -13
Starting timing for: iteration 2
The time was 00:00:07.5153537 for: iteration 2
Training: 3.41103533174678
The learning rate is 2 ** -13
Starting timing for: iteration 3
The time was 00:00:07.4518838 for: iteration 3
Training: 3.38067668991383
The learning rate is 2 ** -13
Starting timing for: iteration 4
The time was 00:00:07.2181451 for: iteration 4
Training: 3.36514445513432
```

Here I actually let the output out.

Now it sure works worse.

11:50am. Let me take a break here."