Let me resume now that I am done with the break, though I have no idea what I will do at this point.
Things really are not going right.

I mean I was prepared for input dependent gradient normalization to fail, but the way it failed is something wholly unexpected to me.

SGD is literally incapable of learning the mean of the outputs passed through ln_relu. A simple linear net just diverges on this very simple task.

And I am supposed to be using this?

Yeah, I know, rather than use SGD for this, I should be using an averaging scheme.

But why not use an averaging scheme to learn a value function instead if SGD is just so messed up?

Maybe I owe an apology to Littman and Isbel. When the averagers section came up during the course, I was literally disgusted at it, but now I am not sure what to believe.

I feel like studying other methods for a while.

Even since I went from KFAC to AC, I've yet to make a single step forward. I do not know how to propagate rewards properly, I do not know how to do pruning and I do not know how to get rid of the need to tune learning rates by hand.

There are so many unanswered questions about deep RL.

If only I could figure out how to deal with reward scaling properly that would be great.

I need to do metalearning. I need to figure out how to reason about differentiable programming. I need to understand what assumptions are going into my program.

Instead of using squared error, why not use Hebbian learning plus EMA updates to weights plus L2 normalization of inputs?

I want to learn a little about model based ML.

If only I could make just a small step forward. If only I could make the linear part completely consistent...

There is something to gradient normalization, but not like this.

Hacks like this never go far. It might have worked for nature, but as a programmer in order to reason about something I need to understand it...

Very well.

What I'll do here is take a break for a while. I was hoping to resume poker by the end of the year, but it is obvious that will not be happening.

I won't happen until I come to an understanding of things.